Simplifying recipe step API for implementers.
