+++ src/Orchard.Web/Modules/Orchard.Recipes/Orchard.Recipes.csproj
@@ -84,18 +84,22 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="RecipeBuilders\ContentRecipeBuilderStep.cs" />
     <Compile Include="RecipeBuilders\RecipeMetadataBuilderStep.cs" />
-    <Compile Include="RecipeBuilders\SiteSettingsBuilderStep.cs" />
-    <Compile Include="RecipeHandlers\CommandRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\DataRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\FeatureRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\MetadataRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\MigrationRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\ModuleRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\SettingsRecipeHandler.cs" />
-    <Compile Include="RecipeHandlers\ThemeRecipeHandler.cs" />
+    <Compile Include="RecipeBuilders\SettingsBuilderStep.cs" />
+    <Compile Include="RecipeExecutionSteps\CommandStep.cs" />
+    <Compile Include="RecipeExecutionSteps\ContentStep.cs" />
+    <Compile Include="RecipeExecutionSteps\FeatureStep.cs" />
+    <Compile Include="RecipeExecutionSteps\ContentSchemaStep.cs" />
+    <Compile Include="RecipeExecutionSteps\MigrationStep.cs" />
+    <Compile Include="RecipeExecutionSteps\ModuleStep.cs" />
+    <Compile Include="RecipeExecutionSteps\SettingsStep.cs" />
+    <Compile Include="RecipeExecutionSteps\ThemeStep.cs" />
+    <Compile Include="RecipeHandlers\RecipeExecutionStepHandler.cs" />
     <Compile Include="Routes.cs" />
     <Compile Include="Services\BuildContext.cs" />
+    <Compile Include="Services\IRecipeExecutionStep.cs" />
     <Compile Include="Services\IRecipeExecutor.cs" />
+    <Compile Include="Services\RecipeExecutionContext.cs" />
+    <Compile Include="Services\RecipeExecutionStep.cs" />
     <Compile Include="Services\RecipeExecutor.cs" />
     <Compile Include="Services\IRecipeBuilder.cs" />
     <Compile Include="Services\RecipeBuilder.cs" />
@@ -152,7 +156,7 @@
     <Content Include="Views\EditorTemplates\ExportSteps\RecipeMetadata.cshtml" />
   </ItemGroup>
   <ItemGroup>
-    <Content Include="Views\EditorTemplates\ExportSteps\SiteSettings.cshtml" />
+    <Content Include="Views\EditorTemplates\ExportSteps\Settings.cshtml" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
