+++ src/Orchard.Web/Modules/Orchard.Recipes/Services/IRecipeExecutionStep.cs
@@ -1,0 +1,6 @@
+﻿namespace Orchard.Recipes.Services {
+    public interface IRecipeExecutionStep : IDependency {
+        string Name { get; }
+        void Execute(RecipeExecutionContext context);
+    }
+}
