+++ src/Orchard.Web/Modules/Orchard.Recipes/Services/RecipeExecutionContext.cs
@@ -1,0 +1,8 @@
+﻿using Orchard.Recipes.Models;
+
+namespace Orchard.Recipes.Services {
+    public class RecipeExecutionContext {
+        public string ExecutionId { get; set; }
+        public RecipeStep RecipeStep { get; set; }
+    }
+}
