+++ src/Orchard.Web/Modules/Orchard.Recipes/Services/RecipeExecutionStep.cs
@@ -1,0 +1,6 @@
+namespace Orchard.Recipes.Services {
+    public abstract class RecipeExecutionStep : Component, IRecipeExecutionStep {
+        public abstract string Name { get; }
+        public abstract void Execute(RecipeExecutionContext context);
+    }
+}
